+++ src/components/Post/elements/ProfileImgAndLikeButton.js
@@ -1,21 +1,20 @@
+import Likes from "./Likes";
+
 import PostContext from "../../../contexts/PostContext";
-import { formattedNumberOfLikes } from "../../../utils/PostsUtils";
 
 import styled from "styled-components";
 import { Link } from "react-router-dom";
-import { AiOutlineHeart, AiTwotoneHeart } from "react-icons/ai";
 import { useContext } from "react";
 
 export default function ProfileImgAndLikeButton() {
-    const { user, likes} = useContext(PostContext);
-    const hasUserLikedThisPost = false;
+    const { user } = useContext(PostContext);
+
     return (
         <Wrapper>
             <Link to={`/user/${user.id}`}>
                 <img src = { user.avatar } alt = {user.username} />
             </Link>
-            {hasUserLikedThisPost ? <LikedHeart /> : <NotLikedHeart />}
-            <p>{ formattedNumberOfLikes(likes.length) }</p>
+            <Likes />
         </Wrapper>
     );
 }
@@ -48,24 +47,4 @@ const Wrapper = styled.div`
             font-size: 11px;
         }
     }
-`
-
-const LikedHeart = styled(AiTwotoneHeart)`
-    font-size: 20px;
-    color: #AC0000;
-    margin-bottom: 6px;
-    @media(max-width: 937px) {
-        font-size: 17px;
-        margin-bottom: 12px;
-    }
-`
-
-const NotLikedHeart = styled(AiOutlineHeart)`
-    font-size: 20px;
-    color: #FFFFFF;
-    margin-bottom: 6px;
-    @media(max-width: 937px) {
-        font-size: 17px;
-        margin-bottom: 12px;
-    }
-`
+`;
